+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1434,7 +1434,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsDatasetsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-02-02T09:27:09.676Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-02-02T10:53:13.684Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1502,7 +1502,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-02-02T09:27:09.675Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-02-02T10:53:13.683Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1570,7 +1570,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-02-02T09:27:09.674Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-02-02T10:53:13.682Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1665,7 +1665,7 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-02T09:26:56.098Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-02T10:53:00.225Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")