+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGETTT0AX159VYG94F98Q8G3"
+        "01KGEZRCYP9WYHEQ5K1X6VTY1K"
     )
     r"""The unique identifier of the data source"""